+++ src/bootstrap.php
@@ -2,14 +2,6 @@
 
 // formatters are reversable // Modifiers are NOT reversable
 
-use Misery\Component\Action\EmptyAction;
-use Misery\Component\Action\FilterAction;
-use Misery\Component\Action\UnsetAction;
-use Misery\Component\Akeneo\Client\HttpWriterFactory;
-use Misery\Component\Converter\Akeneo\Api\Attribute;
-use Misery\Component\Converter\AkeneoCsvHeaderContext;
-use Misery\Component\Converter\AkeneoCsvToStructuredDataConverter;
-
 // Path to your .env file (assuming it's in the project root)
 $envFilePath = __DIR__ . '/../.env';
 
@@ -66,7 +58,7 @@ $converterRegistry->registerAllByName(
         new Misery\Component\Converter\AkeneoCsvHeaderContext()
     ),
     new Misery\Component\Converter\Akeneo\Csv\AttributeOption(),
-    new Misery\Component\Converter\Akeneo\Csv\FamilyVariant(),
+    new \Misery\Component\Converter\Akeneo\Api\FamilyVariant(),
     new Misery\Component\Converter\Akeneo\Csv\ReferenceEntities()
 );
